+++ Brainstem101/BSSectionView.h
@@ -15,13 +15,13 @@
 @property (nonatomic, strong)   CALayer *pathLayer;
 @property (nonatomic, strong)   CALayer *selectionLayer;
 @property (nonatomic, strong)   CALayer *arteryLayer;
-@property (assign)              int sxnNumber;
+@property (assign)              NSInteger sxnNumber;
 @property (assign)              bool isRotated;
 
 - (id)      initWithFrame:(CGRect)frame andSectionNumber:(int) num;
 - (BOOL)    isPoint:(CGPoint)point inPath:(BSStructurePath*) path;
 - (void)    highlightPath:(BSStructurePath *) path;
-- (void)    setNewSectionNumber:(int)num;
+- (void)    setNewSectionNumber:(NSInteger)num;
 
 - (void)    drawPathsForProfile:(NSMutableArray *)inputPaths;
 - (void)    fade;
@@ -33,8 +33,8 @@
 - (void)    rotateViewLeft;
 - (void)    arteryImageNamed:(NSString *)imageName;
 
-+ (CGRect)  getFrameForSection:(int)num;
-+ (CGPoint) getCenterForSection:(int)num;
++ (CGRect)  getFrameForSection:(NSInteger)num;
++ (CGPoint) getCenterForSection:(NSInteger)num;
 + (CGRect)  getFrameForClinicalSection;
 
 @end
